QL-001 Code refactoring and added Class for Login [PG, KM]
