+++ app/build.gradle
@@ -43,4 +43,5 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     androidTestImplementation 'androidx.test.espresso:espresso-intents:3.1.0'
     androidTestImplementation 'androidx.test:rules:1.3.0-beta01'
+    implementation 'com.github.kittinunf.fuel:fuel:2.3.1'
 }
