+++ app/src/androidTest/java/com/tugraz/quizlet/frontend/LoginTest.kt
@@ -1,4 +1,4 @@
-package com.example.quizlet
+package com.tugraz.quizlet.frontend
 
 
 import android.view.View
@@ -12,6 +12,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
 import androidx.test.runner.AndroidJUnit4
+import com.example.quizlet.R
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
