+++ app/src/main/java/com/tugraz/quizlet/frontend/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.example.quizlet
+package com.tugraz.quizlet.frontend
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.example.quizlet.R
 
 
 class MainActivity : AppCompatActivity() {
