+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/orange"
-    tools:context=".MainActivity">
+    tools:context="com.tugraz.quizlet.frontend.MainActivity">
 
     <TextView
         android:id="@+id/textView"
